power_manager: fix linter issues
Fixed various issues reported by cpplint.py when run on
src/platform2/power_manager. A few errors were false-positives
so silenced them by using "// NOLINT"

BUG=None
TEST=FEATURES=test emerge-link platform2
cpplint.py --filter=-build/include_order,+build/include_alpha `find power_manager/* | grep "\.cc\|\.h"`

Reviewed-on: https://chromium-review.googlesource.com/209831
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>

+++ power_manager/powerd/daemon.cc
@@ -131,11 +131,11 @@ std::string GetPowerStatusBatteryDebugString(
       break;
   }
 
-  long rounded_actual = lround(status.battery_percentage);
-  long rounded_display = lround(status.display_battery_percentage);
-  output += base::StringPrintf("%ld%%", rounded_actual);
+  int rounded_actual = lround(status.battery_percentage);
+  int rounded_display = lround(status.display_battery_percentage);
+  output += base::StringPrintf("%d%%", rounded_actual);
   if (rounded_actual != rounded_display)
-    output += base::StringPrintf(" (displayed as %ld%%)", rounded_display);
+    output += base::StringPrintf(" (displayed as %d%%)", rounded_display);
   output += base::StringPrintf(", %.3f/%.3fAh at %.3fA", status.battery_charge,
       status.battery_charge_full, status.battery_current);
 
@@ -224,7 +224,7 @@ int RunSetuidHelper(const std::string& action,
 }  // namespace
 
 // Performs actions requested by |state_controller_|.  The reason that
-// this is a nested class of Daemon rather than just being implented as
+// this is a nested class of Daemon rather than just being implemented as
 // part of Daemon is to avoid method naming conflicts.
 class Daemon::StateControllerDelegate
     : public policy::StateController::Delegate {
